requirements.txt
pandas
pymcel
nbformat
rebound
plotly
matplotlib
pyserial